+++ app/Http/Controllers/BusinessPartnerController.php
@@ -14,6 +14,11 @@ class BusinessPartnerController extends Controller
 
     public function index()
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $filter = new BusinessPartnerFilter(BusinessPartner::query(), session('filters.partners'));
         $businessPartners = $filter->filter()->orderBy('name')->paginate(100);
         return view('business-partner.index', compact('businessPartners'));
@@ -21,11 +26,21 @@ class BusinessPartnerController extends Controller
 
     public function create()
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('business-partner.create');
     }
 
     public function store(Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         BusinessPartner::create($request->all());
@@ -42,11 +57,21 @@ class BusinessPartnerController extends Controller
 
     public function edit(BusinessPartner $business_partner)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('business-partner.edit', compact('business_partner'));
     }
 
     public function update(BusinessPartner $business_partner, Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         $business_partner->update($request->all());
@@ -58,6 +83,11 @@ class BusinessPartnerController extends Controller
 
     public function destroy(BusinessPartner $business_partner)
     {
+        if (\Gate::denies('delete', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $business_partner->delete();
 
         flash('Business partner has been deleted', 'success');
@@ -67,6 +97,11 @@ class BusinessPartnerController extends Controller
 
     public function filter(Request $request)
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $data = $request->only(['name', 'type']);
         \Session::set('filters.partners', $data);
         return \Redirect::back();
