+++ app/Http/Controllers/ProjectController.php
@@ -16,8 +16,6 @@ use PHPExcel;
 
 class ProjectController extends Controller
 {
-
-
     protected $rules = ['name' => 'required'];
 
     public function index()
@@ -29,11 +27,20 @@ class ProjectController extends Controller
 
     public function create()
     {
+        if (!Auth::user()->is_admin) {
+            flash('You are not authorized to do this action');
+            return \Redirect::route('project.index');
+        }
         return view('project.create');
     }
 
     public function store(Request $request)
     {
+        if (!Auth::user()->is_admin) {
+            flash('You are not authorized to do this action');
+            return \Redirect::route('project.index');
+        }
+        
         $this->validate($request, $this->rules);
         $project = Project::create($request->all());
         $project->users()->sync($request->get('users', []));
@@ -45,6 +52,11 @@ class ProjectController extends Controller
 
     public function show(Project $project)
     {
+        if (\Gate::denies('budget', $project) && \Gate::denies('reports', $project)) {
+            flash('You are not authorized to do this action');
+            return \Redirect::route('project.index');
+        }
+
         set_time_limit(1800);
         return view('project.show', compact('project', 'divisions'));
     }
@@ -67,11 +79,22 @@ class ProjectController extends Controller
 
     public function edit(Project $project)
     {
+        if (\Gate::denies('modify', $project)) {
+            flash('You are not authorized to do this action');
+            return \Redirect::route('project.index');
+        }
+
         return view('project.edit', compact('project'));
     }
 
     public function update(Project $project, Request $request)
     {
+
+        if (\Gate::denies('modify', $project)) {
+            flash('You are not authorized to do this action');
+            return \Redirect::route('project.index');
+        }
+
         $this->validate($request, $this->rules);
 
         $project->update($request->all());
@@ -84,6 +107,11 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        if (\Gate::denies('modify', $project)) {
+            flash('You are not authorized to do this action');
+            return \Redirect::route('project.index');
+        }
+
         $project->delete();
 
         flash('Project has been deleted', 'success');
@@ -102,10 +130,12 @@ class ProjectController extends Controller
 
     function costControl(Project $project)
     {
+        if (\Gate::denies('cost_control', $project) && \Gate::denies('reports', $project)) {
+            flash('You are not authorized to do this action');
+            return \Redirect::route('project.index');
+        }
+
         $period = $project->open_period;
         return view('project.cost-control', compact('project', 'period'));
     }
-
-
-
 }
