+++ app/Http/Controllers/ResourcesController.php
@@ -20,7 +20,6 @@ use App\Unit;
 use App\UnitAlias;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Route;
-use phpDocumentor\Reflection\Types\Null_;
 
 class ResourcesController extends Controller
 {
@@ -29,6 +28,11 @@ class ResourcesController extends Controller
 
     public function index()
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $filter = new ResourcesFilter(Resources::query(), session('filters.resources'));
         $resources = $filter->filter()->basic()->orderBy('resource_code')->orderBy('name')->paginate(100);
         return view('resources.index', compact('resources'));
@@ -36,6 +40,17 @@ class ResourcesController extends Controller
 
     public function create()
     {
+        if ($project_id = request('project')) {
+            $project = Project::find($project_id);
+            if (\Gate::denies('resources', $project)) {
+                flash("You don't have access to this page");
+                return \Redirect::to('/');
+            }
+        } elseif (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $units_drop = Unit::options();
         $partners = BusinessPartner::options();
         $resource_types = ResourceType::lists('name', 'id')->all();
@@ -46,7 +61,18 @@ class ResourcesController extends Controller
 
     public function store(Request $request)
     {
-        $request['project_id'] = $request->project;
+        if ($project_id = request('project')) {
+            $project = Project::find($project_id);
+            if (\Gate::denies('resources', $project)) {
+                flash("You don't have access to this page");
+                return \Redirect::to('/');
+            }
+        }
+
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
 
         $this->validate($request, $this->rules);
         if ($request['waste'] <= 1) {
@@ -54,6 +80,7 @@ class ResourcesController extends Controller
         } else {
             $request['waste'] = ($request->waste / 100);
         }
+
         $resource = Resources::create($request->all());
 
         flash('Resource has been saved', 'success');
@@ -66,11 +93,23 @@ class ResourcesController extends Controller
 
     public function show(Resources $resource)
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('resources.show', compact('resource'));
     }
 
     public function edit(Resources $resources)
     {
+
+
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $partners = BusinessPartner::options();
         $resource_types = ResourceType::lists('name', 'id')->all();
         $units_drop = Unit::options();
@@ -81,16 +120,25 @@ class ResourcesController extends Controller
 
     public function update(Resources $resources, Request $request)
     {
-//        $this->validate($request, $this->rules);
+        if ($resources->project_id) {
+            if (\Gate::denies('resources', $resources->project)) {
+                flash("You don't have access to this page");
+                return \Redirect::to('/');
+            }
+        }
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
+        $this->validate($request, $this->rules);
         if ($request['waste'] <= 1) {
             $request['waste'] = $request->waste;
         } else {
             $request['waste'] = ($request->waste / 100);
         }
 
-
         $resources->update($request->all());
-
         $resources->syncCodes($request->get('codes'));
 
         flash('Resource has been saved', 'success');
@@ -102,6 +150,18 @@ class ResourcesController extends Controller
 
     public function destroy(Resources $resources)
     {
+        if ($resources->project_id) {
+            if (\Gate::denies('resources', $resources->project)) {
+                flash("You don't have access to this page");
+                return \Redirect::to('/');
+            }
+        }
+
+        if (\Gate::denies('delete', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $resources->delete();
 
         flash('Resources has been deleted', 'success');
@@ -111,12 +171,22 @@ class ResourcesController extends Controller
 
     function import()
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('resources.import');
     }
 
     function postImport(Request $request)
     {
         $project = $request->project;
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, [
             'file' => 'required|file'//|mimes:xls,xlsx',
         ]);
@@ -143,6 +213,11 @@ class ResourcesController extends Controller
 
     function fixImport($key)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         if (!\Cache::has($key)) {
             flash('Nothing to fix');
             return \Redirect::route('resources.index');
@@ -156,6 +231,11 @@ class ResourcesController extends Controller
 
     function postFixImport(Request $request, $key)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         if (!\Cache::has($key)) {
             flash('Nothing to fix');
             return \Redirect::route('resources.index');
@@ -191,6 +271,11 @@ class ResourcesController extends Controller
 
     function override(Resources $resources, Project $project)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $override = true;
         $overwrote = Resources::version($project->id, $resources->id)->first();
 
@@ -204,6 +289,11 @@ class ResourcesController extends Controller
 
     function postOverride(Resources $resources, Project $project, Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
         $newResource = Resources::version($project->id, $resources->id)->first();
 
@@ -226,6 +316,11 @@ class ResourcesController extends Controller
 
     function filter(Request $request)
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $data = $request->only(['name', 'unit', 'resource_type_id', 'resource_code']);
         \Session::set('filters.resources', $data);
 
@@ -234,6 +329,10 @@ class ResourcesController extends Controller
 
     public function exportResources(Project $project)
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
         $this->dispatch(new ExportResourcesJob($project));
     }
 
@@ -250,11 +349,21 @@ class ResourcesController extends Controller
 
     function importCodes()
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('resources.import-codes');
     }
 
     function postImportCodes(Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, [
             'file' => 'required|file|mimes:xls,xlsx',
         ]);
@@ -270,18 +379,32 @@ class ResourcesController extends Controller
 
     public function exportAllResources()
     {
-        return $this->dispatch(new ExportPublicResourcesJob());
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
 
+        return $this->dispatch(new ExportPublicResourcesJob());
     }
 
     public function modifyAllResources()
     {
-
         return view('resources.modify', ['project_id' => request('project')]);
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
+        return view('resources.modify');
     }
 
     public function postModifyAllResources(Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $project = $request->project;
         $file = $request->file('file');
         $this->dispatch(new ModifyPublicResourcesJob($file, $project));
