+++ app/Http/Controllers/UnitController.php
@@ -14,6 +14,11 @@ class UnitController extends Controller
 
     public function index()
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $filter = new UnitFilter(Unit::query(),session('filters.unit'));
         $units = $filter->filter()->paginate(10);
         return view('unit.index', compact('units'));
@@ -21,11 +26,21 @@ class UnitController extends Controller
 
     public function create()
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         return view('unit.create');
     }
 
     public function store(Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         Unit::create(['type'=>$request->type]);
@@ -42,12 +57,21 @@ class UnitController extends Controller
 
     public function edit(Unit $unit)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
 
         return view('unit.edit', compact('unit'));
     }
 
     public function update(Unit $unit, Request $request)
     {
+        if (\Gate::denies('write', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $this->validate($request, $this->rules);
 
         $unit->update($request->all());
@@ -59,6 +83,11 @@ class UnitController extends Controller
 
     public function destroy(Unit $unit)
     {
+        if (\Gate::denies('delete', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $unit->delete();
 
         flash('Unit has been deleted', 'success');
@@ -68,10 +97,16 @@ class UnitController extends Controller
 
     public function filter(Request $request)
     {
+        if (\Gate::denies('read', 'resources')) {
+            flash("You don't have access to this page");
+            return \Redirect::to('/');
+        }
+
         $data = $request->only('type');
         \Session::set('filters.unit',$data);
         return \Redirect::back();
     }
+
     function wipe(WipeRequest $request)
     {
         \DB::table('units')->delete();
