+++ resources/views/breakdown-template/show.blade.php
@@ -6,17 +6,20 @@
     @else
 
     <form action="{{ route('breakdown-template.destroy', $breakdown_template)}}" class="pull-right" method="post">
-        {{csrf_field()}}
-        {{method_field('delete')}}
 
-        <a href="{{ route('breakdown-template.create', ['activity' => $breakdown_template->std_activity_id])}}"
-           class="btn btn-sm btn-primary">
-            <i class="fa fa-plus"></i> Add template
-        </a>
-        <a href="{{ route('breakdown-template.edit', $breakdown_template)}}" class="btn btn-sm btn-primary">
-            <i class="fa fa-edit"></i> Edit
-        </a>
-        <button class="btn btn-sm btn-warning" type="submit"><i class="fa fa-trash-o"></i> Delete</button>
+        @can('write', 'breakdown-template')
+            <a href="{{ route('breakdown-template.create', ['activity' => $breakdown_template->std_activity_id])}}"
+               class="btn btn-sm btn-primary">
+                <i class="fa fa-plus"></i> Add template
+            </a>
+            <a href="{{ route('breakdown-template.edit', $breakdown_template)}}" class="btn btn-sm btn-primary">
+                <i class="fa fa-edit"></i> Edit
+            </a>
+        @endcan
+        @can('write', 'breakdown-template')
+            {{csrf_field()}} {{method_field('delete')}}
+            <button class="btn btn-sm btn-warning" type="submit"><i class="fa fa-trash-o"></i> Delete</button>
+        @endcan
         <a href="{{ route('std-activity.show', $breakdown_template->activity)}}" class="btn btn-sm btn-default">
             <i class="fa fa-chevron-left"></i> Activity
         </a>
@@ -43,10 +46,12 @@
     </table>
     <h4 class="page-header">Resources</h4>
     <div class="form-group clearfix">
-        <a href="/std-activity-resource/create?template={{$breakdown_template->id}}&&project_id={{request('project_id')}}
-                "
-           class="btn btn-primary btn-sm pull-right"><i class="fa fa-plus-circle"></i>
-            Add Resource</a>
+        @can('write', 'breakdown-template')
+            <a href="/std-activity-resource/create?template={{$breakdown_template->id}}&project_id={{request('project_id')}}"
+               class="btn btn-primary btn-sm pull-right">
+                <i class="fa fa-plus-circle"></i> Add Resource
+            </a>
+        @endcan
     </div>
 
     @if ($breakdown_template->resources->count())
@@ -64,11 +69,17 @@
                     <td class="col-xs-6">{{$resource->resource->name}}</td>
                     <td class="col-xs-3">{{$resource->equation}}</td>
                     <td class="col-xs-3">
-                        {{Form::model($resource, ['route' => ['std-activity-resource.destroy', $resource], 'method' => 'delete'])}}
-                        <a href="{{route('std-activity-resource.edit', $resource)}}" class="btn btn-primary btn-sm"><i
-                                    class="fa fa-edit"></i> Edit</a>
-                        <button class="btn btn-warning btn-sm"><i class="fa fa-trash"></i> Remove</button>
-                        {{Form::close()}}
+                        @can('write', 'breakdown-template')
+                            {{Form::model($resource, ['route' => ['std-activity-resource.destroy', $resource], 'method' => 'delete'])}}
+                            <a href="{{route('std-activity-resource.edit', $resource)}}" class="btn btn-primary btn-sm">
+                                <i class="fa fa-edit"></i> Edit
+                            </a>
+
+                            @can('delete', 'breakdown-template')
+                                <button class="btn btn-warning btn-sm"><i class="fa fa-trash"></i> Remove</button>
+                            @endcan
+                            {{Form::close()}}
+                        @endcan
                     </td>
                 </tr>
             @endforeach
@@ -77,5 +88,4 @@
     @else
         <div class="alert alert-info"><i class="fa fa-info-circle"></i> No resources found</div>
     @endif
-
 @stop
